requirements2.txt
streamlit
requests
python-dotenv
google-generativeai
